Chore: modularize TokenStore
