+++ lib/token-store/index.js
@@ -8,10 +8,12 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const lodash = require("lodash");
+const cursors = require("./cursors");
+const ForwardTokenCursor = require("./forward-token-cursor");
+const PaddedTokenCursor = require("./padded-token-cursor");
 
 //------------------------------------------------------------------------------
-// Implementation
+// Helpers
 //------------------------------------------------------------------------------
 
 const PUBLIC_METHODS = Object.freeze([
@@ -37,474 +39,6 @@ const PUBLIC_METHODS = Object.freeze([
     "getTokenOrCommentBefore",
     "getTokenOrCommentAfter"
 ]);
-
-/**
- * Gets `token.range[0]` from the given token.
- *
- * @param {Node|Token|Comment} token - The token to get.
- * @returns {number} The start location.
- * @private
- */
-function getStartLocation(token) {
-    return token.range[0];
-}
-
-/**
- * Binary-searches the index of the first token which is after the given location.
- * If it was not found, this returns `tokens.length`.
- *
- * @param {(Token|Comment)[]} tokens - It searches the token in this list.
- * @param {number} location - The location to search.
- * @returns {number} The found index or `tokens.length`.
- * @private
- */
-function search(tokens, location) {
-    return lodash.sortedIndexBy(tokens, { range: [location] }, getStartLocation);
-}
-
-/**
- * Gets the index of the `startLoc` in `tokens`.
- * `startLoc` can be the value of `node.range[1]`, so this checks about `startLoc - 1` as well.
- *
- * @param {(Token|Comment)[]} tokens - The tokens to find an index.
- * @param {Object} indexMap - The map from locations to indices.
- * @param {number} startLoc - The location to get an index.
- * @returns {number} The index.
- * @private
- */
-function getFirstIndex(tokens, indexMap, startLoc) {
-    if (startLoc in indexMap) {
-        return indexMap[startLoc];
-    }
-    if ((startLoc - 1) in indexMap) {
-        return indexMap[startLoc - 1] + 1;
-    }
-    return 0;
-}
-
-/**
- * Gets the index of the `endLoc` in `tokens`.
- * The information of end locations are recorded at `endLoc - 1` in `indexMap`, so this checks about `endLoc - 1` as well.
- *
- * @param {(Token|Comment)[]} tokens - The tokens to find an index.
- * @param {Object} indexMap - The map from locations to indices.
- * @param {number} endLoc - The location to get an index.
- * @returns {number} The index.
- * @private
- */
-function getLastIndex(tokens, indexMap, endLoc) {
-    if (endLoc in indexMap) {
-        return indexMap[endLoc] - 1;
-    }
-    if ((endLoc - 1) in indexMap) {
-        return indexMap[endLoc - 1];
-    }
-    return tokens.length - 1;
-}
-
-/**
- * The abstract class about cursors which iterate tokens.
- *
- * This class has 2 abstract methods.
- *
- * - `current: Token | Comment | null` ... The current token.
- * - `moveNext(): boolean` ... Moves this cursor to the next token. If the next token didn't exist, it returns `false`.
- *
- * There are the following known sub classes.
- *
- * - ForwardTokenCursor .......... The cursor which iterates tokens only.
- * - BackwardTokenCursor ......... The cursor which iterates tokens only in reverse.
- * - ForwardTokenCommentCursor ... The cursor which iterates tokens and comments.
- * - BackwardTokenCommentCursor .. The cursor which iterates tokens and comments in reverse.
- * - DecorativeCursor
- *     - FilterCursor ............ The cursor which ignores the specified tokens.
- *     - SkipCursor .............. The cursor which ignores the first few tokens.
- *     - LimitCursor ............. The cursor which limits the count of tokens.
- *
- * @private
- */
-class Cursor {
-
-    /**
-     * Initializes this cursor.
-     */
-    constructor() {
-        this.current = null;
-    }
-
-    /**
-     * Gets the first token.
-     * This consumes this cursor.
-     * @returns {Token|Comment} The first token or null.
-     */
-    getOneToken() {
-        return this.moveNext() ? this.current : null;
-    }
-
-    /**
-     * Gets the first tokens.
-     * This consumes this cursor.
-     * @returns {(Token|Comment)[]} All tokens.
-     */
-    getAllTokens() {
-        const tokens = [];
-
-        while (this.moveNext()) {
-            tokens.push(this.current);
-        }
-
-        return tokens;
-    }
-
-    /**
-     * Moves this cursor to the next token.
-     * @returns {boolean} `true` if the next token exists.
-     * @abstract
-     */
-    /* istanbul ignore next */
-    moveNext() { // eslint-disable-line class-methods-use-this
-        throw new Error("Not implemented.");
-    }
-}
-
-/**
- * The cursor which iterates tokens only.
- * @private
- */
-class ForwardTokenCursor extends Cursor {
-
-    /**
-     * Initializes this cursor.
-     * @param {Token[]} tokens - The array of tokens.
-     * @param {Comment[]} comments - The array of comments.
-     * @param {Object} indexMap - The map from locations to indices in `tokens`.
-     * @param {number} startLoc - The start location of the iteration range.
-     * @param {number} endLoc - The end location of the iteration range.
-     */
-    constructor(tokens, comments, indexMap, startLoc, endLoc) {
-        super();
-        this.tokens = tokens;
-        this.index = getFirstIndex(tokens, indexMap, startLoc);
-        this.indexEnd = getLastIndex(tokens, indexMap, endLoc);
-    }
-
-    /** @inheritdoc */
-    moveNext() {
-        if (this.index <= this.indexEnd) {
-            this.current = this.tokens[this.index];
-            this.index += 1;
-            return true;
-        }
-        return false;
-    }
-
-    //
-    // Shorthand for performance.
-    //
-
-    /** @inheritdoc */
-    getOneToken() {
-        return (this.index <= this.indexEnd) ? this.tokens[this.index] : null;
-    }
-
-    /** @inheritdoc */
-    getAllTokens() {
-        return this.tokens.slice(this.index, this.indexEnd + 1);
-    }
-}
-
-/**
- * The cursor which iterates tokens only in reverse.
- * @private
- */
-class BackwardTokenCursor extends Cursor {
-
-    /**
-     * Initializes this cursor.
-     * @param {Token[]} tokens - The array of tokens.
-     * @param {Comment[]} comments - The array of comments.
-     * @param {Object} indexMap - The map from locations to indices in `tokens`.
-     * @param {number} startLoc - The start location of the iteration range.
-     * @param {number} endLoc - The end location of the iteration range.
-     */
-    constructor(tokens, comments, indexMap, startLoc, endLoc) {
-        super();
-        this.tokens = tokens;
-        this.index = getLastIndex(tokens, indexMap, endLoc);
-        this.indexEnd = getFirstIndex(tokens, indexMap, startLoc);
-    }
-
-    /** @inheritdoc */
-    moveNext() {
-        if (this.index >= this.indexEnd) {
-            this.current = this.tokens[this.index];
-            this.index -= 1;
-            return true;
-        }
-        return false;
-    }
-
-    //
-    // Shorthand for performance.
-    //
-
-    /** @inheritdoc */
-    getOneToken() {
-        return (this.index >= this.indexEnd) ? this.tokens[this.index] : null;
-    }
-}
-
-/**
- * The cursor which iterates tokens and comments.
- * @private
- */
-class ForwardTokenCommentCursor extends Cursor {
-
-    /**
-     * Initializes this cursor.
-     * @param {Token[]} tokens - The array of tokens.
-     * @param {Comment[]} comments - The array of comments.
-     * @param {Object} indexMap - The map from locations to indices in `tokens`.
-     * @param {number} startLoc - The start location of the iteration range.
-     * @param {number} endLoc - The end location of the iteration range.
-     */
-    constructor(tokens, comments, indexMap, startLoc, endLoc) {
-        super();
-        this.tokens = tokens;
-        this.comments = comments;
-        this.tokenIndex = getFirstIndex(tokens, indexMap, startLoc);
-        this.commentIndex = search(comments, startLoc);
-        this.border = endLoc;
-    }
-
-    /** @inheritdoc */
-    moveNext() {
-        const token = (this.tokenIndex < this.tokens.length) ? this.tokens[this.tokenIndex] : null;
-        const comment = (this.commentIndex < this.comments.length) ? this.comments[this.commentIndex] : null;
-
-        if (token && (!comment || token.range[0] < comment.range[0])) {
-            this.current = token;
-            this.tokenIndex += 1;
-        } else if (comment) {
-            this.current = comment;
-            this.commentIndex += 1;
-        } else {
-            this.current = null;
-        }
-
-        return Boolean(this.current) && (this.border === -1 || this.current.range[1] <= this.border);
-    }
-}
-
-/**
- * The cursor which iterates tokens and comments in reverse.
- * @private
- */
-class BackwardTokenCommentCursor extends Cursor {
-
-    /**
-     * Initializes this cursor.
-     * @param {Token[]} tokens - The array of tokens.
-     * @param {Comment[]} comments - The array of comments.
-     * @param {Object} indexMap - The map from locations to indices in `tokens`.
-     * @param {number} startLoc - The start location of the iteration range.
-     * @param {number} endLoc - The end location of the iteration range.
-     */
-    constructor(tokens, comments, indexMap, startLoc, endLoc) {
-        super();
-        this.tokens = tokens;
-        this.comments = comments;
-        this.tokenIndex = getLastIndex(tokens, indexMap, endLoc);
-        this.commentIndex = search(comments, endLoc) - 1;
-        this.border = startLoc;
-    }
-
-    /** @inheritdoc */
-    moveNext() {
-        const token = (this.tokenIndex >= 0) ? this.tokens[this.tokenIndex] : null;
-        const comment = (this.commentIndex >= 0) ? this.comments[this.commentIndex] : null;
-
-        if (token && (!comment || token.range[1] > comment.range[1])) {
-            this.current = token;
-            this.tokenIndex -= 1;
-        } else if (comment) {
-            this.current = comment;
-            this.commentIndex -= 1;
-        } else {
-            this.current = null;
-        }
-
-        return Boolean(this.current) && (this.border === -1 || this.current.range[0] >= this.border);
-    }
-}
-
-/**
- * The cursor which iterates tokens only.
- * This is for the backward compatibility of padding options.
- * @private
- */
-class PaddedTokenCursor extends ForwardTokenCursor {
-
-    /**
-     * Initializes this cursor.
-     * @param {Token[]} tokens - The array of tokens.
-     * @param {Comment[]} comments - The array of comments.
-     * @param {Object} indexMap - The map from locations to indices in `tokens`.
-     * @param {number} startLoc - The start location of the iteration range.
-     * @param {number} endLoc - The end location of the iteration range.
-     * @param {number} beforeCount - The number of tokens this cursor iterates before start.
-     * @param {number} afterCount - The number of tokens this cursor iterates after end.
-     */
-    constructor(tokens, comments, indexMap, startLoc, endLoc, beforeCount, afterCount) {
-        super(tokens, comments, indexMap, startLoc, endLoc);
-        this.index = Math.max(0, this.index - beforeCount);
-        this.indexEnd = Math.min(tokens.length - 1, this.indexEnd + afterCount);
-    }
-}
-
-/**
- * The abstract class about cursors which manipulate another cursor.
- * @private
- */
-class DecorativeCursor extends Cursor {
-
-    /**
-     * Initializes this cursor.
-     * @param {Cursor} cursor - The cursor to be decorated.
-     */
-    constructor(cursor) {
-        super();
-        this.cursor = cursor;
-    }
-
-    /** @inheritdoc */
-    moveNext() {
-        const retv = this.cursor.moveNext();
-
-        this.current = this.cursor.current;
-
-        return retv;
-    }
-}
-
-/**
- * The decorative cursor which ignores specified tokens.
- * @private
- */
-class FilterCursor extends DecorativeCursor {
-
-    /**
-     * Initializes this cursor.
-     * @param {Cursor} cursor - The cursor to be decorated.
-     * @param {Function} predicate - The predicate function to decide tokens this cursor iterates.
-     */
-    constructor(cursor, predicate) {
-        super(cursor);
-        this.predicate = predicate;
-    }
-
-    /** @inheritdoc */
-    moveNext() {
-        const predicate = this.predicate;
-
-        while (super.moveNext()) {
-            if (predicate(this.current)) {
-                return true;
-            }
-        }
-        return false;
-    }
-}
-
-/**
- * The decorative cursor which ignores the first few tokens.
- * @private
- */
-class SkipCursor extends DecorativeCursor {
-
-    /**
-     * Initializes this cursor.
-     * @param {Cursor} cursor - The cursor to be decorated.
-     * @param {number} count - The count of tokens this cursor skips.
-     */
-    constructor(cursor, count) {
-        super(cursor);
-        this.count = count;
-    }
-
-    /** @inheritdoc */
-    moveNext() {
-        while (this.count > 0) {
-            this.count -= 1;
-            if (!super.moveNext()) {
-                return false;
-            }
-        }
-        return super.moveNext();
-    }
-}
-
-/**
- * The decorative cursor which limits the number of tokens.
- * @private
- */
-class LimitCursor extends DecorativeCursor {
-
-    /**
-     * Initializes this cursor.
-     * @param {Cursor} cursor - The cursor to be decorated.
-     * @param {number} count - The count of tokens this cursor iterates.
-     */
-    constructor(cursor, count) {
-        super(cursor);
-        this.count = count;
-    }
-
-    /** @inheritdoc */
-    moveNext() {
-        if (this.count > 0) {
-            this.count -= 1;
-            return super.moveNext();
-        }
-        return false;
-    }
-}
-
-/**
- * The cursor pair of tokens-only and tokens/comments.
- * @private
- */
-class CursorPair {
-
-    /**
-     * Initializes this cursor.
-     * @param {Function} TokenCursor - The class of the cursor which iterates tokens only.
-     * @param {Function} TokenCommentCursor - The class of the cursor which iterates the mix of tokens and comments.
-     */
-    constructor(TokenCursor, TokenCommentCursor) {
-        this.TokenCursor = TokenCursor;
-        this.TokenCommentCursor = TokenCommentCursor;
-    }
-
-    /**
-     * Creates the cursor iterates tokens with normalized options.
-     *
-     * @param {Token[]} tokens - The array of tokens.
-     * @param {Comment[]} comments - The array of comments.
-     * @param {Object} indexMap - The map from locations to indices in `tokens`.
-     * @param {number} startLoc - The start location of the iteration range.
-     * @param {number} endLoc - The end location of the iteration range.
-     * @param {boolean} includeComments - The flag to iterate comments as well.
-     * @returns {Cursor} The created cursor.
-     * @private
-     */
-    createCursor(tokens, comments, indexMap, startLoc, endLoc, includeComments) {
-        return new (includeComments ? this.TokenCommentCursor : this.TokenCursor)(tokens, comments, indexMap, startLoc, endLoc);
-    }
-}
-
-const forwardCursors = Object.freeze(new CursorPair(ForwardTokenCursor, ForwardTokenCommentCursor));
-const backwardCursors = Object.freeze(new CursorPair(BackwardTokenCursor, BackwardTokenCommentCursor));
 
 /**
  * Creates the map from locations to indices in `tokens`.
@@ -544,41 +78,9 @@ function createIndexMap(tokens, comments) {
 }
 
 /**
- * Creates the cursor iterates tokens with normalized options.
- *
- * @param {CursorPair} cursors - The cursor pair to initialize base cursor.
- * @param {Token[]} tokens - The array of tokens.
- * @param {Comment[]} comments - The array of comments.
- * @param {Object} indexMap - The map from locations to indices in `tokens`.
- * @param {number} startLoc - The start location of the iteration range.
- * @param {number} endLoc - The end location of the iteration range.
- * @param {boolean} includeComments - The flag to iterate comments as well.
- * @param {Function|null} filter - The predicate function to choose tokens.
- * @param {number} skip - The count of tokens the cursor skips.
- * @param {number} count - The maximum count of tokens the cursor iterates. Zero is no iteration for backward compatibility.
- * @returns {Cursor} The created cursor.
- * @private
- */
-function createCursor(cursors, tokens, comments, indexMap, startLoc, endLoc, includeComments, filter, skip, count) {
-    let cursor = cursors.createCursor(tokens, comments, indexMap, startLoc, endLoc, includeComments);
-
-    if (filter) {
-        cursor = new FilterCursor(cursor, filter);
-    }
-    if (skip >= 1) {
-        cursor = new SkipCursor(cursor, skip);
-    }
-    if (count >= 0) {
-        cursor = new LimitCursor(cursor, count);
-    }
-
-    return cursor;
-}
-
-/**
  * Creates the cursor iterates tokens with options.
  *
- * @param {CursorPair} cursors - The cursor pair to initialize base cursor.
+ * @param {CursorFactory} factory - The cursor factory to initialize cursor.
  * @param {Token[]} tokens - The array of tokens.
  * @param {Comment[]} comments - The array of comments.
  * @param {Object} indexMap - The map from locations to indices in `tokens`.
@@ -591,7 +93,7 @@ function createCursor(cursors, tokens, comments, indexMap, startLoc, endLoc, inc
  * @returns {Cursor} The created cursor.
  * @private
  */
-function createCursorWithSkip(cursors, tokens, comments, indexMap, startLoc, endLoc, opts) {
+function createCursorWithSkip(factory, tokens, comments, indexMap, startLoc, endLoc, opts) {
     let includeComments = false;
     let skip = 0;
     let filter = null;
@@ -606,13 +108,13 @@ function createCursorWithSkip(cursors, tokens, comments, indexMap, startLoc, end
         filter = opts.filter || null;
     }
 
-    return createCursor(cursors, tokens, comments, indexMap, startLoc, endLoc, includeComments, filter, skip, -1);
+    return factory.createCursor(tokens, comments, indexMap, startLoc, endLoc, includeComments, filter, skip, -1);
 }
 
 /**
  * Creates the cursor iterates tokens with options.
  *
- * @param {CursorPair} cursors - The cursor pair to initialize base cursor.
+ * @param {CursorFactory} factory - The cursor factory to initialize cursor.
  * @param {Token[]} tokens - The array of tokens.
  * @param {Comment[]} comments - The array of comments.
  * @param {Object} indexMap - The map from locations to indices in `tokens`.
@@ -625,7 +127,7 @@ function createCursorWithSkip(cursors, tokens, comments, indexMap, startLoc, end
  * @returns {Cursor} The created cursor.
  * @private
  */
-function createCursorWithCount(cursors, tokens, comments, indexMap, startLoc, endLoc, opts) {
+function createCursorWithCount(factory, tokens, comments, indexMap, startLoc, endLoc, opts) {
     let includeComments = false;
     let count = 0;
     let filter = null;
@@ -641,7 +143,7 @@ function createCursorWithCount(cursors, tokens, comments, indexMap, startLoc, en
         filter = opts.filter || null;
     }
 
-    return createCursor(cursors, tokens, comments, indexMap, startLoc, endLoc, includeComments, filter, 0, count);
+    return factory.createCursor(tokens, comments, indexMap, startLoc, endLoc, includeComments, filter, 0, count);
 }
 
 /**
@@ -680,20 +182,35 @@ function createCursorWithPadding(tokens, comments, indexMap, startLoc, endLoc, b
     if (typeof beforeCount === "number" || typeof beforeCount === "undefined") {
         return new PaddedTokenCursor(tokens, comments, indexMap, startLoc, endLoc, beforeCount | 0, afterCount | 0);
     }
-    return createCursorWithCount(forwardCursors, tokens, comments, indexMap, startLoc, endLoc, beforeCount);
+    return createCursorWithCount(cursors.forward, tokens, comments, indexMap, startLoc, endLoc, beforeCount);
 }
 
-class TokenStore {
+//------------------------------------------------------------------------------
+// Exports
+//------------------------------------------------------------------------------
+
+/**
+ * The token store.
+ *
+ * This class provides methods to get tokens by locations as fast as possible.
+ * The methods are a part of public API, so we should be careful if it changes this class.
+ *
+ * People can get tokens in O(1) by the hash map which is mapping from the location of tokens/comments to tokens.
+ * Also people can get a mix of tokens and comments in O(log k), the k is the number of comments.
+ * Assuming that comments to be much fewer than tokens, this does not make hash map from token's locations to comments to reduce memory cost.
+ * This uses binary-searching instead for comments.
+ */
+module.exports = class TokenStore {
 
     /**
      * Initializes this token store.
      *
-     * ※ `comments` needs cloning for backward compatibility.
+     * ※ `comments` needs to be cloned for backward compatibility.
      * After this initialization, ESLint removes a shebang's comment from `comments`.
-     * However, so far we had been concatenating 'tokens' and 'comments',
+     * However, so far we had been concatenating 'tokens' and 'comments' before,
      * so the shebang's comment had remained in the concatenated array.
-     * As the result, both `getTokenOrCommentAfter` and `getTokenOrCommentBefore`
-     * methods had been returning the shebang's comment.
+     * As a result, both the result of `getTokenOrCommentAfter` and `getTokenOrCommentBefore`
+     * methods had included the shebang's comment.
      * And some rules depends on this behavior.
      *
      * @param {Token[]} tokens - The array of tokens.
@@ -701,7 +218,7 @@ class TokenStore {
      */
     constructor(tokens, comments) {
         this.tokens = tokens;
-        this.comments = comments.slice(0); // Needs cloning for backward compatibility!
+        this.comments = comments.slice(0);
         this.indexMap = createIndexMap(tokens, comments);
     }
 
@@ -734,7 +251,7 @@ class TokenStore {
      */
     getFirstToken(node, options) {
         return createCursorWithSkip(
-            forwardCursors,
+            cursors.forward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -755,7 +272,7 @@ class TokenStore {
      */
     getLastToken(node, options) {
         return createCursorWithSkip(
-            backwardCursors,
+            cursors.backward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -776,7 +293,7 @@ class TokenStore {
      */
     getTokenBefore(node, options) {
         return createCursorWithSkip(
-            backwardCursors,
+            cursors.backward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -797,7 +314,7 @@ class TokenStore {
      */
     getTokenAfter(node, options) {
         return createCursorWithSkip(
-            forwardCursors,
+            cursors.forward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -819,7 +336,7 @@ class TokenStore {
      */
     getFirstTokenBetween(left, right, options) {
         return createCursorWithSkip(
-            forwardCursors,
+            cursors.forward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -841,7 +358,7 @@ class TokenStore {
      */
     getLastTokenBetween(left, right, options) {
         return createCursorWithSkip(
-            backwardCursors,
+            cursors.backward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -892,7 +409,7 @@ class TokenStore {
      */
     getFirstTokens(node, options) {
         return createCursorWithCount(
-            forwardCursors,
+            cursors.forward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -913,7 +430,7 @@ class TokenStore {
      */
     getLastTokens(node, options) {
         return createCursorWithCount(
-            backwardCursors,
+            cursors.backward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -934,7 +451,7 @@ class TokenStore {
      */
     getTokensBefore(node, options) {
         return createCursorWithCount(
-            backwardCursors,
+            cursors.backward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -955,7 +472,7 @@ class TokenStore {
      */
     getTokensAfter(node, options) {
         return createCursorWithCount(
-            forwardCursors,
+            cursors.forward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -977,7 +494,7 @@ class TokenStore {
      */
     getFirstTokensBetween(left, right, options) {
         return createCursorWithCount(
-            forwardCursors,
+            cursors.forward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -999,7 +516,7 @@ class TokenStore {
      */
     getLastTokensBetween(left, right, options) {
         return createCursorWithCount(
-            backwardCursors,
+            cursors.backward,
             this.tokens,
             this.comments,
             this.indexMap,
@@ -1065,11 +582,6 @@ class TokenStore {
             padding
         ).getAllTokens();
     }
-}
+};
 
-//------------------------------------------------------------------------------
-// Exports
-//------------------------------------------------------------------------------
-
-module.exports = TokenStore;
 module.exports.PUBLIC_METHODS = PUBLIC_METHODS;
