+++ lib/util/source-code.js
@@ -8,7 +8,7 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const TokenStore = require("../token-store.js"),
+const TokenStore = require("../token-store"),
     Traverser = require("./traverser");
 
 //------------------------------------------------------------------------------
